Add javaGetter as maintained API
